requirements.txt
beanie
quart
quart-cors
python-dotenv
aiohttp
pydantic[email]
requests
Flask-Discord
binary-reader
pillow
matplotlib
pycountry
fuzzy-search
humanize